Added a 'no trailer' option
What Was The Problem:
-----------------------------------

When csvdiff is used as part of a pipeline of applications 
the trailer (Totals etc) can cause issues when a subsequent 
application uses go-langs own CSV parser. The reason for 
this is the go-lang parser assumes the number of columns is 
always the same.


What Did I Do To Solve This:
-----------------------------------
I Added a 't' option to the configuration parameters and 
when added to an executing command the trailer is not 
displayed.

What Do I Verify This:
-----------------------------------
After building the application type:

    csvdiff --help

And you should see:

    -t   No trailer

When you do a diff with the -t option then no tailer should
be seen.